WEEK_5_Project.py/requirements.txt
cffi
cryptography
pycparser
PyMySQL
python-dotenv
six
tabulate